requirements.txt
python-twitter
requests
selenium
beautifulsoup4

